move cheatables to cheatables directory
