+++ protostar/cairo_testing/cairo_test_execution_state.py
@@ -17,7 +17,9 @@ from starkware.storage.storage import FactFetchingContext
 from typing_extensions import Self
 
 from protostar.compiler import ProjectCompiler
-from protostar.cheatable_starknet.cheatable_cached_state import CheatableCachedState
+from protostar.cheatable_starknet.cheatables.cheatable_cached_state import (
+    CheatableCachedState,
+)
 from protostar.testing.stopwatch import Stopwatch
 from protostar.testing.test_config import TestConfig
 from protostar.testing.test_context import TestContext
