+++ protostar/cairo_testing/cheatcode_factories/cairo_test_cheatcode_factory.py
@@ -1,6 +1,8 @@
 from typing import List
 
-from protostar.cheatable_starknet.cheatable_cached_state import CheatableCachedState
+from protostar.cheatable_starknet.cheatables.cheatable_cached_state import (
+    CheatableCachedState,
+)
 from protostar.cheatable_starknet.cheatcodes.prank_cairo_cheatcode import (
     PrankCairoCheatcode,
 )
