+++ protostar/cheatable_starknet/cheaters/block_info.py
@@ -6,7 +6,7 @@ from starkware.starknet.business_logic.state.state_api_objects import BlockInfo
 from protostar.starknet.address import Address
 
 if TYPE_CHECKING:
-    from protostar.cheatable_starknet.cheatable_cached_state import (
+    from protostar.cheatable_starknet.cheatables.cheatable_cached_state import (
         CheatableCachedState,
     )
