+++ protostar/cheatable_starknet/cheaters/contracts.py
@@ -30,7 +30,7 @@ from protostar.contract_types import (
     DeclaredContract,
     DeployedContract,
 )
-from protostar.cheatable_starknet.cheatable_entry_point import (
+from protostar.cheatable_starknet.cheatables.cheatable_entry_point import (
     CheatableExecuteEntryPoint,
 )
 from protostar.starknet.types import ClassHashType
@@ -43,7 +43,9 @@ from protostar.starknet.data_transformer import (
 )
 
 if TYPE_CHECKING:
-    from protostar.cheatable_starknet.cheatable_cached_state import CheatableCachedState
+    from protostar.cheatable_starknet.cheatables.cheatable_cached_state import (
+        CheatableCachedState,
+    )
 
     from . import CairoCheaters
